refactor(eval): address Harrison's review feedback
- move renderJsonTemplate out of shared src/io into core/eval/invokeDataset
- invokeRuntime: raw TypeError/Error -> InputValidationError/RuntimeInvokeResponseError
- wire error causes in template + dataset JSON parse
- invokeDataset: enrich per-example invoke failure instead of log+rethrow
- simulate: bubble Ctrl-C cancellation (telemetry) instead of quiet return; clarify --description help; TODO(#1986) shared abort helper
- drop type-guaranteed 'no leak' test; keep AbortSignal wiring in composition test
- trim stale/redundant comments (runtime.tsx, invokeRuntime DTO note, TurnResult)
